Fixed AsYouType functionality
The AsYouType class had a problem because of how it handled removing a char. If you would remove a char then continue to add chars after the remove, the formatting will fail to update. This was because we were doing a sub string on the formatted char when we should have been storing all the chars and using the google-libphonenumber built in clear then just rebuild the number. 

I have not tested this inside the library. I fixed my code using this technique then modified this code on Github. I just wanted to get a fix in for future people trying to use this function.